requirements.txt
ujson == 5.9.0
pyserial